rename: 네이밍 index로 수정(#18)
